requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-jazzmin==2.6.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djangrestframework==0.0.1
drf-yasg==1.21.7
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
multidict==6.0.4
packaging==23.2
peppercorn==0.6
phonenumbers==8.13.24
Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.12.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
timedelta==2020.12.3
twilio==8.10.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
yarl==1.9.2
